utils/util.py: Fix a bug where the std. dev. was calculated with the mean field in voxels
trainer/trainer.py: Fix a bug where the std. dev. was calculated with the mean field in voxels

logger/logger.py: Fix a bug where the std. dev. was calculated with the mean field in voxels
utils/util.py: Fix a bug where the std. dev. was calculated with the mean field in voxels