trainer/trainer.py: Log the displacement mean and std. dev. to tensorboard and save the displacement corresp. to the mean of the variational posterior
trainer/trainer.py: Log the displacement mean and std. dev. to tensorboard and save the displacement corresp. to the mean of the variational posterior

logger/logger.py: Save the displacement corresp. to the mean of the variational posterior
logger/visualization.py: Log the displacement mean and std. dev. to tensorboard
utils/util.py: Refactor code